Add files via upload
This is a neural network having an input layer, a hidden layer with 2 units and an output layer.
Note: This function is only written to make the students understand how neural networks work. The implementation in packages is quite difference and is optimal in comparison to this code so do not use this code if you need optimal or more accurate results. 
Play with the code as you like